Get MakeAccessibleSVG partly working for geom_lines including multiples
MakeAccessibleSVG.ggplot now handles plots with a single geom_line (more
or less). Because the whole line is a single polyline, can't highlight
segments visually, but have included a work-around so we can still
describe them individually in text.
Have started to handle multiple layers -- only for lines currently, and
only if all layers are lines