Separate whisker preprocessing; print as side-effect; allow threshhod param
Fields only needed in the structure due to the limitations of whisker are
now added in a separate preprocessing step.
VI now prints as a side-effect, and print.VIgraph now correctly returns
the object invisibly.
Threshold parameter is now correctly set up to be passed in to VI.
Incomplete MakeAccessible code for handling panels has been tidied up.