Introduce list separator text.  Include README describing VIgg object
We nnow can print lists with grammatical separators (commas and the word
"and" between the last 2 items).  Axis tick labels are also in a proper
list now (not just a vector) so they can be nicely printed.
Also, list the number of panels at the start for faceted graphs.
Have introduced README.md file in whisker folder that describes the
VIgg object that the template has to work with.