0.33.4 Move BrowseSVG to whiskers (moustache) templates.
As can be seen this decreases the code quite a bit and makes it much simplier. This also goes in line with VI and describe working with moustache templates rather then simple cat function calls. The cacc.js file is also embedded into the template so the result is one *.html file rather than 2 (one html and one js).

This change will also make future expansion to the html and styling less stressful.